+++ src/horsebox/main.py
@@ -249,6 +249,7 @@ def __inspect_cmd(**kwargs: Any) -> None:
 @click.option('--count', type=bool, is_flag=True, default=False, help='Count the number of items found.')
 @click.option('--top', type=bool, is_flag=True, default=False, help='Get the top keywords of the result.')
 @click.option('--source', 'fields', type=str, multiple=True, help='Fields of the indexed documents to output.')
+@click.option('--sort', 'sort_field', type=str, help='Field to sort by the result.')
 @click.option(
     '--explain',
     type=bool,